services/ui/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.2
